Download And Delete Torrent Files
This is the main meat of the downloading logic.
Torrents now have a process for downloading the files recursivly from
the server to the local machine, and then deleteing the torrent and
cleaning up the remote server files, all at the same time.

The main function also now loops forever with a configurable delay.
This lets the scripts function as intended, checking the status of
torrents after a period. (default 30 mins)

New Processes
 - download_files
 - download_files_into_folder
 - stop_tracking_torrent
 - delte_torrent_only
 - delete_torrent_and_files
Other changes
 - torrents can remove themselves from the watching list
 - various XMLRPC functions to support the new processes
 - remove the URL print on startup
 - changed the planned name of delete_torrent to
   delete_torrent_only for clarity
 - main function loops forever
 - download thread has special case for deleting instead
   (I do not trust paramiko to be thread safe)
 - feed updates on print if there were new links
 - download thread has its state saved when the program
   is told to shutdown (leaves partial files)
 - added tags into the download queue
   (support for the download_files processing)
 - download thread will resume when program is restarted
 - downloads are now in its own python thread
 - download destinations can recursivly make folders