Fix Processing Steps Escaped Character
When an escaped character was proccesed in the processing steps,
the index was not advanced and so broke the splitting of processing
steps. This manifested as the closing paren of the step being added
to the name of the next step being processed.

Sinply remove the continue to allow the increment to happen