Bugfix: RuntimeError when expanding new torrents
The code that expanded new torrents would increase the number of
keys in the dictionary. This is intentional, but a standard for loop
will error because the dict generator will be invalid when new keys
are added. Instead we simply make a copy of the keys and itterate
through the copy.