Add string modifications to replace substrings
It is handy sometimes to replace parts of a sting with other things
and so i have added a string mod to do so. Also as a convenience
(and the reason i have the replament mod) there is a mod to
sanitize a string to be usable for a folder/file name.
- add string mod replace()
- add string mod sanitizeFilename()
- also make the torrent update less picky so debugging is easier