Bugfix + Stop Torrent Variable Expansion
Why was i even bothering with expanding torrent variables?
I had some thought that if i made the torernt sections "fat" then you could safely modify the
rss_feeds.conf and everything would finish just fine. But really if someone is
removing a group or something from the rss_feeds.conf file, then they should not be surprised
if something breaks because it depends on it.

Why did i remove this functionality?
Because in the processing steps, some variables are made by certain steps, and if you reference
these new variables (like regex_matched) then you must put an empty useless variable of that
same name in the torrent/group/default realm just so it doesn't complain, which is obviously nonsense.
Also, it is easier to fix something that a current torrent being processed ran into if you only have to change
the group/defaults instead of dealing with a copy in every single torrent because it was wrong.

Specific Changes for removing the variable expansion:
 - when expanding string variables, do not copy that variable into the torrent
 - in expand_new_torrent_object we no longer parse for every referenced variable

Bugfixes:
 - regex parsing did not have a way to report that it failed to match
 - Argument validation error prints for branch_if_vars_equal and branch_if_values_equal had the wrong function names referenced
    - i should probably change the error reporting to use the function name instea dof being hard coded
 - get_file_info syntax was wrong for argument unpacking
 - get_file_info now strips the leading period on the extension
 - populate_next_file_info now passes the filepath argument correctly now