Improved new hash detection
Previous hash grabbing when adding a URL was using the assumption that the
latest added URL would be the last has in the returned list.
When adding many items at once, this assumption is sometimes violated.
Instead the list of hashes is searched specificly for a new hash.

Also, added in a couple sanity checks to the hash selection.