More aggressive saving of torrent status
If you are adding a lot of torrents (20+ in my experiance)
then it may have rtorrent choke. If you kill and restart the automator
then it will have lost all progress of the added torrents and the user
has a lot of manual cleanup that they must do to fix things.
It is better to just save a few extra times and have it gracefully
recover. As such i removed the nearly useless prints saying it saved.