IP Ban workaround
The host i use has started pushing jackett and such, and these make a
lot of requests and so has gotten the server banned from nyaa.si
So to get around the ban of rTorrent not being able to directly grab the
torrent from the URL, simply download it locally and then send it base64
encoded to rTorrent.

Also apparently fix the ratio number being read. I thought this
used to work? Maybe it is a recent change or maybe it has always been
broken?