FileIO Cleanup for job starting
There were 3 places that were doing the same work with nearly identical
formatting to start a job in the threadpool.
Instead move this same work to where the job wrapper is and just start
it there. There is no reason to have a wrapper factory since it is only
ever used for the threadpool jobs.