SFTPDownload handles directories
Major feature update to recusivly download all files at the given remote
location. This can greatly simplify things so that the calling code
doesn't need to iterate through all the files themselves and we also can
give a total job job size/percentage. It is also going to be faster as
we can reuse the same SFTP connection instead of opening and closing the
connecting for every file.

 - Add a new verbose flag for getProcessingStatus
 - SFTPDownload recursivly downloads files from the remote location
 - SFTPDownload makes any local directories required to download
 - rename sftpdownload.py to just sftp.py
 - Major overhaul of the processing counters and reporting