Add group_name when expanding torrents
This allows someone to delete the parent RssFeed item without the
torrent breaking. Previously it relied on finding the group from the
feed. This still does not fully solve dependency on the parent for
things such as grabbing season numbers or regex overrides, but if a
torrent is sitting thre waiting for a ratio, then there is almost never
a need for additional interaction with the feed besides the group_name.